+++ spec/helper/stub_network.rb
@@ -2,7 +2,7 @@
 
 class StubNetwork
   class << self
-    def stub!
+    def stub!(os_type = 'ubuntu')
       unless Kanrisuru::Remote::Host.instance_methods(false).include?(:execute_with_retries_alias)
         Kanrisuru::Remote::Host.class_eval do
           alias_method :execute_with_retries_alias, :execute_with_retries
@@ -18,16 +18,16 @@ class StubNetwork
       unless Kanrisuru::Remote::Os.instance_methods(false).include?(:initialize_alias)
         Kanrisuru::Remote::Os.class_eval do
           alias_method :initialize_alias, :initialize 
-          def initialize(host)
+          define_method :initialize do |host|
             @host = host
 
-            @kernel_name       = 'Linux'
-            @kernel_version    = '#91-Ubuntu SMP Thu Jul 15 19:09:17 UTC 2021'
-            @operating_system  = 'GNU/Linux'
-            @hardware_platform = 'x86_64'
-            @processor         = 'x86_64'
-            @release           = 'ubuntu'
-            @version           = 20.0
+            @kernel_name       = StubNetwork.send(:os_defaults, os_type, :kernel_name)
+            @kernel_version    = StubNetwork.send(:os_defaults, os_type, :kernel_version)
+            @operating_system  = StubNetwork.send(:os_defaults, os_type, :operating_system)
+            @hardware_platform = StubNetwork.send(:os_defaults, os_type, :hardware_platform)
+            @processor         = StubNetwork.send(:os_defaults, os_type, :processor)
+            @release           = StubNetwork.send(:os_defaults, os_type, :release)
+            @version           = StubNetwork.send(:os_defaults, os_type, :version)
           end
         end
       end
@@ -57,6 +57,37 @@ class StubNetwork
       Kanrisuru::Result.class_eval do
         alias_method :initialize, :initialize_alias 
       end
+    end
+
+    private
+
+    def os_defaults(name, property)
+      name = name.to_sym
+
+      defaults = {
+        ubuntu: {
+          kernel_name: 'Linux',
+          kernel_version: '#91-Ubuntu SMP Thu Jul 15 19:09:17 UTC 2021',
+          operating_system: 'GNU/Linux',
+          hardware_platform: 'x86_64',
+          processor: 'x86_64',
+          release: 'ubuntu',
+          version: 20.0
+        },
+        centos: {
+          kernel_name: 'Linux',
+          kernel_version: '"#1 SMP Wed Jul 21 11:57:15 UTC 2021"',
+          operating_system: 'GNU/Linux',
+          hardware_platform: 'x86_64',
+          processor: 'x86_64',
+          release: 'centos',
+          version: 7.0
+        }
+      }
+
+
+      defaults[name].key?(property) ?
+        defaults[name][property] : nil
     end
 
   end